refactor(team-builder): extract ITEM/ABIL/NAT/TERA cells + FormCells wrapper
Final atomic-cell extraction. Each loadout field gets its own component
under lanes/identity/cells/, taking variant: "row" | "grid" to switch
between compact-mode (.formRow chrome) and hero-mode (.heroFormCell
chrome). Each cell owns its own picker-open useState.

Adds FormCells, a thin wrapper that renders the four cells in order so
layout components don't have to repeat the boilerplate. Both branches
of IdentityLaneReal now render <FormCells variant="row" /> or
<FormCells variant="grid" /> instead of inline picker setup.

Deletes IdentityAbilityRow (logic absorbed into AbilityCell) and the
four picker-open useStates from IdentityLaneReal. After this commit,
identity-lane.tsx imports no *Picker components — they all live in
cells.

Sets up step 6 (split into three layout components) where each layout
just composes the cells in different arrangements.
